edit main page and add devise settings

+++ app/controllers/api/v1/articles_controller.rb
@@ -2,7 +2,6 @@ class Api::V1::ArticlesController < ApiController
   before_action :authenticate_user, except: [:index, :show]
   before_action :authorize_user, except: [:index, :show, :create]
   
-  
   def index    
     render json: Article.all
   end
@@ -24,24 +23,26 @@ class Api::V1::ArticlesController < ApiController
     end       
   end 
 
-  def update
+  # def update
 
-    article = Article.find(params[:id])
+  #   article = Article.find(params[:id])
 
-    if article.update(article_params)
-      render json: article
-    else 
-      render json: {error: article.errors.messages}
-    end 
+  #   if article.update(article_params)
+  #     render json: article
+  #   else 
+  #     render json: {error: article.errors.messages}
+  #   end 
 
-  end 
+  # end 
 
 
   def destroy 
     
-    article = Article.find(params[:id])   
-    article.destroy
-    render json: current_user.articles
+    article = Article.find(params[:id])  
+    if article.user = current_user 
+      article.destroy
+      render json: current_user.articles
+    end 
   end 
 
 
@@ -50,6 +51,7 @@ class Api::V1::ArticlesController < ApiController
   def article_params
     params.require(:article).permit(:title,:description)
   end 
+
 
   def authorize_user
     if !user_signed_in? || !(current_user.role == "admin")
